work work

+++ src/lngs/Entity.go
@@ -44,6 +44,12 @@ type Entity struct {
 }
 
 func (self *Entity) SetClient(client *GameClient) {
+	if client == self.client {
+		return // client not changed at all
+	}
+
+	old_client := self.client
+
 	if self.client != nil {
 		self.client.DestroyEntity(self.id)
 
@@ -57,6 +63,14 @@ func (self *Entity) SetClient(client *GameClient) {
 		self.client.CreateEntity(self.id, self.GetBehaviorName())
 		self.client.BecomePlayer(self.id)
 	}
+
+	if old_client != nil && self.client == nil {
+		self.onLoseClient(old_client)
+	} else if old_client == nil && self.client != nil {
+		self.onGetNewClient()
+	} else {
+		self.onChangeClient(old_client)
+	}
 }
 
 func newEntity(id string) *Entity {
@@ -64,6 +78,10 @@ func newEntity(id string) *Entity {
 		id = NewEntityId()
 	}
 	return &Entity{id, nil, noneBehavior, GetCommandQueue(id)}
+}
+
+func (self *Entity) Id() string {
+	return self.id
 }
 
 func (self *Entity) GetBehaviorName() string {
@@ -97,27 +115,12 @@ func (self *Entity) OnReceiveMessage(msg Message) {
 	}
 
 	ARGS := msg["ARGS"].([]interface{})
-	targetEntity.OnCallMethod(self, M, ARGS)
+	targetEntity.CallMethod(M, ARGS...)
 }
 
-func (self *Entity) OnCallMethod(caller *Entity, methodname string, args []interface{}) {
-	log.Printf("%s calling %s.%s", caller, self, methodname)
-
-	method := self.behavior.MethodByName(methodname)
-	log.Printf("method %v, total methods %d", method, self.behavior.NumMethod())
-	in := make([]reflect.Value, len(args)+1)
-	in[0] = reflect.ValueOf(self)
-	for i, arg := range args {
-		in[i+1] = reflect.ValueOf(arg)
-	}
-	// methodType := method.Type()
-	// numArguments := methodType.NumIn()
-	// for argIndex := 0; argIndex < numArguments; argIndex++ {
-	// 	var argType reflect.Type = methodType.In(argIndex)
-	// 	log.Println("arg type", argIndex, argType)
-	// 	in[argIndex] = convertType(in[argIndex], argType)
-	// }
-	method.Call(in)
+func (self *Entity) CallMethod(methodname string, args ...interface{}) {
+	log.Printf("Entity method: %s.%s", self, methodname)
+	self.callBehaviorMethod(methodname, args...)
 }
 
 func (self *Entity) CallClient(method string, args ...interface{}) {
@@ -222,6 +225,38 @@ func (self *Entity) SendCommand(targetid string, cmdName string, cmdData interfa
 		cmdData,
 	}
 	PostCommandQueue(targetid, &cmd)
+}
+
+func (self *Entity) callBehaviorMethod(methodname string, args ...interface{}) bool {
+	method := self.behavior.MethodByName(methodname)
+	if !method.IsValid() {
+		log.Printf("Entity %s: method %s not found\n", self, methodname)
+		return false
+	}
+
+	in := make([]reflect.Value, len(args)+1)
+	in[0] = reflect.ValueOf(self)
+
+	for i, arg := range args {
+		in[i+1] = reflect.ValueOf(arg)
+	}
+	method.Call(in)
+	return true
+}
+
+func (self *Entity) onGetNewClient() {
+	// get new client: self.client
+	self.callBehaviorMethod("OnGetNewClient")
+}
+
+func (self *Entity) onLoseClient(old_client *GameClient) {
+	// lose client: old_client
+	self.callBehaviorMethod("OnLoseClient", old_client)
+}
+
+func (self *Entity) onChangeClient(old_client *GameClient) {
+	// client changed from old_client to self.client
+	self.callBehaviorMethod("OnChangeClient", old_client)
 }
 
 // func convertType(val reflect.Value, targetType reflect.Type) reflect.Value {
